Refactor `refreshTasks()` by creating fetch-tasks.ts
